Fixed: Server-side buffer overflow when trying to transmit a "huge" InFine script to clients. Improved: Added some rudimentary mod authoring, debug aids to InFine scripts. If a referenced graphic is missing, it will now be reported in the console but the script will continue to play (objects will using said graphic will be drawn without a texture).

+++ doomsday/engine/portable/src/net_main.c
@@ -284,9 +284,9 @@ boolean Net_GetPacket(void)
 /**
  * This is the public interface of the message sender.
  */
-void Net_SendPacket(int to_player, int type, void *data, int length)
+void Net_SendPacket(int to_player, int type, void *data, size_t length)
 {
-    int         flags = 0;
+    int                 flags = 0;
 
     // What kind of delivery to use?
     if(to_player & DDSP_CONFIRM)
